modify based on comments
